+++ src/main/kotlin/com/openlattice/chronicle/services/upload/AppDataUploadService.kt
@@ -22,7 +22,9 @@ import com.openlattice.chronicle.util.ChronicleServerUtil
 import com.geekbeast.postgres.PostgresColumnDefinition
 import com.geekbeast.postgres.PostgresDatatype
 import com.geekbeast.postgres.PostgresTableDefinition
+import com.openlattice.chronicle.sensorkit.SensorDataSample
 import com.openlattice.chronicle.storage.PostgresDataTables
+import com.openlattice.chronicle.storage.RedshiftDataTables
 import com.zaxxer.hikari.HikariDataSource
 import org.apache.commons.lang3.RandomStringUtils
 import org.apache.olingo.commons.api.edm.FullQualifiedName
@@ -46,6 +48,27 @@ class AppDataUploadService(
     private val organizationSettingsManager: OrganizationSettingsManager
 ) : AppDataUploadManager {
     private val logger = LoggerFactory.getLogger(AppDataUploadService::class.java)
+
+    companion object {
+        private val SENSOR_DATA_COLUMNS = RedshiftDataTables.IOS_SENSOR.columns.joinToString(",") { it.name }
+
+        /**
+         * PreparedStatement bind order
+         * 1) organizationId,
+         * 2) studyId,
+         * 3) participantId,
+         * 4) id
+         * 5) sensorType
+         * 6) dateRecorded
+         * 7) duration
+         * 8) timezone
+         * 9) sensorData,
+         */
+        private val INSERT_SENSOR_DATA_SQL = """
+            INSERT INTO ${RedshiftDataTables.IOS_SENSOR.name}($SENSOR_DATA_COLUMNS) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?) 
+            ON CONFLICT DO NOTHING
+        """.trimIndent()
+    }
 
     private fun getTruncatedDateTimeHelper(dateTime: String?, chronoUnit: ChronoUnit): String? {
         return if (dateTime == null) {
@@ -189,6 +212,65 @@ class AppDataUploadService(
         }
     }
 
+    override fun uploadSensorData(studyId: UUID, participantId: String, sourceDeviceId: String, data: List<SensorDataSample>): Int {
+
+        StopWatch(
+                log = "logging ${data.size} entries for ${ChronicleServerUtil.STUDY_PARTICIPANT_DATASOURCE}",
+                level = Level.INFO,
+                logger = logger,
+                data.size,
+                studyId,
+                participantId,
+                sourceDeviceId
+        ).use {
+            try {
+                val status = enrollmentManager.getParticipationStatus(studyId, participantId)
+                if (ParticipationStatus.NOT_ENROLLED == status) {
+                    logger.warn(
+                            "participant is not enrolled, ignoring sensor data upload" + ChronicleServerUtil.STUDY_PARTICIPANT_DATASOURCE,
+                            studyId,
+                            participantId,
+                            sourceDeviceId
+                    )
+                    return 0
+                }
+                val deviceEnrolled = enrollmentManager.isKnownDatasource(studyId, participantId, sourceDeviceId)
+
+                if (!deviceEnrolled) {
+                    logger.error(
+                            "data source not found, ignoring sensor data upload" + ChronicleServerUtil.STUDY_PARTICIPANT_DATASOURCE,
+                            studyId,
+                            participantId,
+                            sourceDeviceId
+                    )
+                    return 0
+                }
+
+                logger.info(
+                        "attempting to upload sensor data" + ChronicleServerUtil.STUDY_PARTICIPANT_DATASOURCE,
+                        studyId,
+                        participantId,
+                        sourceDeviceId
+                )
+                val written = writeSensorDataToRedshift( studyId, participantId, data)
+                if (written != data.size) {
+                    logger.warn("Wrote $written entities, but wrote ${data.size}")
+                }
+
+                return written
+            } catch (ex: Exception) {
+                logger.error(
+                        "error writing sensor data" + ChronicleServerUtil.STUDY_PARTICIPANT_DATASOURCE,
+                        studyId,
+                        participantId,
+                        sourceDeviceId,
+                        ex
+                )
+                return 0
+            }
+        }
+    }
+
     private fun filter(
         organizationId: UUID, mappedData: Sequence<Map<String, UsageEventColumn>>
     ): Sequence<Map<String, UsageEventColumn>> {
@@ -296,9 +378,39 @@ class AppDataUploadService(
         )
     }
 
+    private fun writeSensorDataToRedshift(
+            studyId: UUID,
+            participantId: String,
+            data: List<SensorDataSample>
+    ) : Int {
 
-    companion object {
-        private const val APP_USAGE_FREQUENCY = "appUsageFrequency"
+        val (_, hds) = storageResolver.resolveAndGetFlavor(studyId)
+
+        return hds.connection.use { connection ->
+            try {
+                val wc = connection.prepareStatement(INSERT_SENSOR_DATA_SQL).use { ps ->
+                    // TODO: make sure this works~!~
+                    ps.setString(2, studyId.toString())
+                    ps.setString(3, participantId)
+
+                    data.forEach { sample ->
+                        ps.setString(4, sample.id.toString())
+                        ps.setString(5, sample.sensor)
+                        ps.setObject(6, sample.dateRecorded)
+                        ps.setDouble(7, sample.duration)
+                        ps.setString(8, sample.timezone)
+                        ps.setString(9, sample.data)
+//                        ps.setString(10, sample.device)
+
+                        ps.addBatch()
+                    }
+                    ps.executeBatch().sum()
+                }
+                return@use wc
+            } catch (ex: Exception) {
+                throw ex
+            }
+        }
     }
 }
